requirements.txt
pandas>=0.25
searchtweets-v2>=1.1.1